make a basic pagination
in this commit the css file has been reformated, the indentation has
been reduced to 2 spaces, to be more compliant to usually style guide in
place for css files and enforce by prettier.

second thing has been to reoganize the files, article file deleted to
create bogus file in root content as the section will be located here.
the bogus file is a example of a recipe file.

note: it's not obvious but the parameter "paginate_by" must be set in
the section file (_index.md) to make pagination work